experiments/countdown-workflow/pyproject.toml
[project]
name = "countdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.26.0",
    "moviepy>=2.2.1",
    "yt-dlp>=2025.6.30",
]
